scripts/report_performance.py
from __future__ import annotations

from pathlib import Path
import sys

PROJECT_ROOT = Path(__file__).resolve().parents[1]
if str(PROJECT_ROOT) not in sys.path:
    sys.path.insert(0, str(PROJECT_ROOT))

import json

from dotenv import load_dotenv

from value_engine.roi.tracker import build_performance_summary


def main() -> None:
    load_dotenv()
    summary = build_performance_summary()
    print(json.dumps(summary, ensure_ascii=False, indent=2))


if __name__ == "__main__":
    main()
